requirements.txt
flask
wtforms
python-dotenv
pymongo[srv]
gunicorn